refactor IDM xml configuration into config module and move IDM tests into idm/tests
