+++ config/src/main/java/org/picketlink/config/idm/resolver/PropertyResolver.java
@@ -20,7 +20,7 @@
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
 
-package org.picketlink.idm.config.internal.resolver;
+package org.picketlink.config.idm.resolver;
 
 /**
  * Interface for resolve property value from configuration (usually XML configuration) to real property value, which will be used
