+++ idm/impl/pom.xml
@@ -73,19 +73,10 @@
             <version>${project.version}</version>
           </dependency>
 
-          <dependency>
-            <groupId>org.picketlink</groupId>
-            <artifactId>picketlink-config</artifactId>
-            <version>${project.version}</version>
-          </dependency>
 	<dependency>
             <groupId>org.picketlink</groupId>
             <artifactId>picketlink-idm-api</artifactId>
         </dependency>
-    <!--dependency>
-      <groupId>org.picketlink</groupId>
-      <artifactId>picketlink-config</artifactId>
-    </dependency-->
 		<dependency>
 			<groupId>junit</groupId>
 			<artifactId>junit</artifactId>
