+++ idm/tests/src/test/java/org/picketlink/test/idm/TestLifecycle.java
