+++ idm/tests/src/test/java/org/picketlink/test/idm/usecases/ApplicationRegistrationTestCase.java
