+++ idm/tests/src/test/resources/META-INF/beans.xml
