+++ idm/tests/src/test/resources/config/embedded-file-config.xml
